Update the extractor package and QAs structures, Extract and Run methods sign
Modify the name of the files in the extractor package to be more appropriate.
Question and answer are stored together in a single object and a slice of it is
used to store QAs.
Extract method reflects the changes applied to QAs, as well as the quiz struct.
Run uses a timer to show the available time to a user in order to complete a quiz.